+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/content/nixie/foundation/GenericNixieDisplayBlockEntity.java
@@ -1,10 +1,12 @@
 package com.kipti.bnb.content.nixie.foundation;
 
 import com.kipti.bnb.CreateBitsnBobs;
-import com.kipti.bnb.content.nixie.nixie_board.NixieBoardBlock;
+import com.kipti.bnb.content.nixie.nixie_board.NixieBoardBlockNixie;
+import com.kipti.bnb.mixin_accessor.DynamicComponentMigrator;
 import com.kipti.bnb.registry.BnbBlocks;
 import com.simibubi.create.foundation.blockEntity.SmartBlockEntity;
 import com.simibubi.create.foundation.blockEntity.behaviour.BlockEntityBehaviour;
+import com.simibubi.create.foundation.utility.DynamicComponent;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.HolderLookup;
@@ -17,29 +19,64 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
 
+    private static final int MAX_CHARS_IN_BLOCK = 3;
+
     private static final Logger log = LoggerFactory.getLogger(GenericNixieDisplayBlockEntity.class);
-    protected String currentTextTop = "";
-    protected String currentTextBottom = "";
+
+    private Optional<DynamicComponent> customTextTop = Optional.empty();
+    private Optional<DynamicComponent> customTextBottom = Optional.empty();
+    private int customTextStart = 0;
+
+    private String renderedTextTop = "";
+    private String renderedTextBottom = "";
+
     protected ConfigurableDisplayOptions currentDisplayOption = ConfigurableDisplayOptions.NONE;
 
-    public GenericNixieDisplayBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state) {
+    public GenericNixieDisplayBlockEntity(final BlockEntityType<?> type, final BlockPos pos, final BlockState state) {
         super(type, pos, state);
     }
 
-    public void inheritDataFrom(GenericNixieDisplayBlockEntity be) {
-        this.currentTextTop = be.currentTextTop;
-        this.currentTextBottom = be.currentTextBottom;
+    @Override
+    public void initialize() {
+        if (level.isClientSide)
+            updateDisplayedStrings();
+    }
+
+    public void inheritDataFrom(final GenericNixieDisplayBlockEntity be) {
+        this.customTextTop = be.customTextTop;
+        this.customTextBottom = be.customTextBottom;
+        this.customTextStart = be.customTextStart;
+
         this.currentDisplayOption = be.currentDisplayOption;
+
         notifyUpdate();
     }
 
     public ConfigurableDisplayOptions getCurrentDisplayOption() {
         return currentDisplayOption;
+    }
+
+    public char getRenderedText(int line, int width) {
+        if (line == 0) {
+            if (width < renderedTextTop.length()) {
+                return renderedTextTop.charAt(width);
+            }
+        } else if (line == 1) {
+            if (width < renderedTextBottom.length()) {
+                return renderedTextBottom.charAt(width);
+            }
+        }
+        return ' ';
+    }
+
+    public void setPositionOffset(final int consumedCharsOnRow) {
+        this.customTextStart = consumedCharsOnRow;
     }
 
     public enum ConfigurableDisplayOptions {
@@ -52,7 +89,7 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         public final int width;
         public final int lines;
 
-        ConfigurableDisplayOptions(int width, int lines, Supplier<ConfigurableDisplayOptionTransform> renderTransform) {
+        ConfigurableDisplayOptions(final int width, final int lines, final Supplier<ConfigurableDisplayOptionTransform> renderTransform) {
             this.renderTransform = renderTransform;
             this.width = width;
             this.lines = lines;
@@ -64,15 +101,15 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         return ((IGenericNixieDisplayBlock) getBlockState().getBlock()).getPossibleDisplayOptions();
     }
 
-    public void applyToEachElementOfThisStructure(Consumer<GenericNixieDisplayBlockEntity> consumer) {
-        GenericNixieDisplayBlockEntity controller = findControllerBlockEntity();
-        Direction facing = getBlockState().getValue(DoubleOrientedDisplayBlock.FACING);
-        Direction orientation = getBlockState().getValue(DoubleOrientedDisplayBlock.ORIENTATION);
-        Direction right = DoubleOrientedDirections.getLeft(facing, orientation).getOpposite();
+    public void applyToEachElementOfThisStructure(final Consumer<GenericNixieDisplayBlockEntity> consumer) {
+        final GenericNixieDisplayBlockEntity controller = findControllerBlockEntity();
+        final Direction facing = getBlockState().getValue(GenericNixieDisplayBlock.FACING);
+        final Direction orientation = getBlockState().getValue(GenericNixieDisplayBlock.ORIENTATION);
+        final Direction right = DoubleOrientedDirections.getLeft(facing, orientation).getOpposite();
         BlockPos currentPos = controller.getBlockPos();
         for (int i = 0; i < 100; i++) {
-            BlockEntity blockEntity = controller.level.getBlockEntity(currentPos);
-            if (blockEntity instanceof GenericNixieDisplayBlockEntity display && areStatesComprableForConnection(controller.getBlockState(), display.getBlockState())) {
+            final BlockEntity blockEntity = controller.level.getBlockEntity(currentPos);
+            if (blockEntity instanceof final GenericNixieDisplayBlockEntity display && areStatesComprableForConnection(controller.getBlockState(), display.getBlockState())) {
                 consumer.accept(display);
             } else {
                 break; // No more display found
@@ -81,31 +118,20 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         }
     }
 
-    public void setDisplayOption(ConfigurableDisplayOptions option) {
+    public void setDisplayOption(final ConfigurableDisplayOptions option) {
         if (currentDisplayOption == option) {
             return; // No change
         }
         currentDisplayOption = option;
-        //Apply clipping to the text
-        EndClipping endClipping = getEndClipping();
-        if (endClipping != EndClipping.NONE) {
-            //Ensure any text inside the clipping region is space
-            String topRaw = currentTextTop.trim();
-            String bottomRaw = currentTextBottom.trim();
-            int displayedCharacterWidth = calculateDisplayedCharacterWidth();
-            currentTextTop = " ".repeat(endClipping.left) + (topRaw.isEmpty() ? "" : topRaw.substring(0, Math.min(displayedCharacterWidth, topRaw.length())));
-            currentTextBottom = " ".repeat(endClipping.left) + (bottomRaw.isEmpty() ? "" : bottomRaw.substring(0, Math.min(displayedCharacterWidth, bottomRaw.length())));
-        } else if (currentDisplayOption == ConfigurableDisplayOptions.NONE) {
-            currentTextTop = currentTextTop.trim();
-            currentTextBottom = currentTextBottom.trim();
-        }
+        if (level.isClientSide)
+            updateDisplayedStrings();
         notifyUpdate();
     }
 
     /*
      * Apparently, comprable isn't a word, i do not care, I will use it anyway.
      * */
-    public static boolean areStatesComprableForConnection(BlockState state1, BlockState state2) {
+    public static boolean areStatesComprableForConnection(final BlockState state1, final BlockState state2) {
         if (state1 == null || state2 == null) {
             return false;
         }
@@ -116,48 +142,111 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         final boolean stateTwoIsTube = BnbBlocks.LARGE_NIXIE_TUBE.is(state2.getBlock()) || BnbBlocks.DYED_LARGE_NIXIE_TUBE.contains(state2.getBlock());
 
         if (!(stateOneIsBoard && stateTwoIsBoard) && !(stateOneIsTube && stateTwoIsTube)) return false;
-        if (state1.getValue(DoubleOrientedDisplayBlock.FACING) != state2.getValue(DoubleOrientedDisplayBlock.FACING)) {
+        if (state1.getValue(GenericNixieDisplayBlock.FACING) != state2.getValue(GenericNixieDisplayBlock.FACING)) {
             return false;
         }
-        return state1.getValue(DoubleOrientedDisplayBlock.ORIENTATION) == state2.getValue(DoubleOrientedDisplayBlock.ORIENTATION);
+        return state1.getValue(GenericNixieDisplayBlock.ORIENTATION) == state2.getValue(GenericNixieDisplayBlock.ORIENTATION);
     }
 
     @Override
-    public void addBehaviours(List<BlockEntityBehaviour> behaviours) {
+    public void addBehaviours(final List<BlockEntityBehaviour> behaviours) {
 
     }
 
     @Override
-    protected void write(CompoundTag tag, HolderLookup.Provider registries, boolean clientPacket) {
+    protected void write(final CompoundTag tag, final HolderLookup.Provider registries, final boolean clientPacket) {
         super.write(tag, registries, clientPacket);
-        tag.putString("CurrentTextTop", currentTextTop);
-        tag.putString("CurrentTextBottom", currentTextBottom);
+
+        if (customTextTop.isPresent()) {
+            final CompoundTag componentTag = new CompoundTag();
+            customTextTop.get()
+                    .write(componentTag, registries);
+            tag.put("CustomTextComponentTop", componentTag);
+        }
+        if (customTextBottom.isPresent()) {
+            final CompoundTag componentTag = new CompoundTag();
+            customTextBottom.get()
+                    .write(componentTag, registries);
+            tag.put("CustomTextComponentBottom", componentTag);
+        }
+
+        tag.putInt("CustomTextIndex", customTextStart);
         tag.putString("CurrentDisplayOption", currentDisplayOption.name());
     }
 
     @Override
-    protected void read(CompoundTag tag, HolderLookup.Provider registries, boolean clientPacket) {
+    protected void read(final CompoundTag tag, final HolderLookup.Provider registries, final boolean clientPacket) {
         super.read(tag, registries, clientPacket);
-        currentTextTop = tag.getString("CurrentTextTop");
-        currentTextBottom = tag.getString("CurrentTextBottom");
-        String optionName = tag.getString("CurrentDisplayOption");
+
+        if (tag.contains("CustomTextComponentTop")) {
+            final DynamicComponent component = customTextTop.orElseGet(DynamicComponent::new);
+            final CompoundTag componentTag = tag.getCompound("CustomTextComponentTop");
+            component.read(worldPosition, componentTag, registries);
+            customTextTop = Optional.of(component);
+        }
+
+        if (tag.contains("CustomTextComponentBottom")) {
+            final DynamicComponent component = customTextBottom.orElseGet(DynamicComponent::new);
+            final CompoundTag componentTag = tag.getCompound("CustomTextComponentBottom");
+            component.read(worldPosition, componentTag, registries);
+            customTextBottom = Optional.of(component);
+        }
+
+        customTextStart = tag.getInt("CustomTextIndex");
+
+        final String optionName = tag.getString("CurrentDisplayOption");
         try {
             currentDisplayOption = ConfigurableDisplayOptions.valueOf(optionName);
-        } catch (IllegalArgumentException e) {
+        } catch (final IllegalArgumentException e) {
             log.warn("Invalid display option '{}' for block entity at {}. Defaulting to NONE.", optionName, getBlockPos());
             currentDisplayOption = ConfigurableDisplayOptions.NONE;
         }
+
+        //Migration: Check for old tag names
+        //TODO: remove migration by release 1.0.1
+        if (tag.contains("CurrentTextTop")) {
+            log.info("Migrating old Nixie display data for block entity at {}", getBlockPos());
+            String oldTextTop = tag.getString("CurrentTextTop");
+            if (!oldTextTop.isEmpty()) {
+                final EndClipping clipping = getEndClipping();
+                final DynamicComponent component = new DynamicComponent();
+                ((DynamicComponentMigrator) component).bits_n_bobs$setValueToLiteral(oldTextTop, registries);
+                customTextTop = Optional.of(component);
+                customTextStart = clipping.left;
+            } else {
+                customTextTop = Optional.empty();
+                customTextStart = 0;
+            }
+        }
+
+        if (tag.contains("CurrentTextBottom")) {
+            log.info("Migrating old Nixie display data for block entity at {}", getBlockPos());
+            final String oldTextBottom = tag.getString("CurrentTextBottom");
+            if (!oldTextBottom.isEmpty()) {
+                final DynamicComponent component = new DynamicComponent();
+                ((DynamicComponentMigrator) component).bits_n_bobs$setValueToLiteral(oldTextBottom, registries);
+                customTextBottom = Optional.of(component);
+                final EndClipping clipping = getEndClipping();
+                customTextStart = clipping.left;
+            } else {
+                customTextBottom = Optional.empty();
+                customTextStart = 0;
+            }
+        }
+
+        if (clientPacket || isVirtual())
+            updateDisplayedStrings();
     }
 
     public @NotNull GenericNixieDisplayBlockEntity findControllerBlockEntity() {
-        Direction facing = getBlockState().getValue(DoubleOrientedDisplayBlock.FACING);
-        Direction orientation = getBlockState().getValue(DoubleOrientedDisplayBlock.ORIENTATION);
-        Direction left = DoubleOrientedDirections.getLeft(facing, orientation);
+        final Direction facing = getBlockState().getValue(GenericNixieDisplayBlock.FACING);
+        final Direction orientation = getBlockState().getValue(GenericNixieDisplayBlock.ORIENTATION);
+        final Direction left = DoubleOrientedDirections.getLeft(facing, orientation);
         BlockPos leftPos = getBlockPos().relative(left);
         GenericNixieDisplayBlockEntity lastDisplay = this;
         for (int i = 0; i < 100; i++) {
-            BlockEntity blockEntity = level.getBlockEntity(leftPos);
-            if (blockEntity instanceof GenericNixieDisplayBlockEntity display && areStatesComprableForConnection(getBlockState(), display.getBlockState())) {
+            final BlockEntity blockEntity = level.getBlockEntity(leftPos);
+            if (blockEntity instanceof final GenericNixieDisplayBlockEntity display && areStatesComprableForConnection(getBlockState(), display.getBlockState())) {
                 lastDisplay = display;
             } else {
                 break; // No more display found
@@ -167,45 +256,39 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         return lastDisplay;
     }
 
-    public void applyTextToDisplay(String tagElement, int line) {
-        EndClipping endClipping = getEndClipping();
-        int consumptionWidth = calculateDisplayedCharacterWidth();
-        String consumed = " ".repeat(endClipping.left) + (tagElement.isEmpty() ? "" : tagElement.substring(0, Math.min(consumptionWidth, tagElement.length())));
-        if (line == 0) {
-            this.currentTextTop = consumed;
-        } else if (line == 1) {
-            this.currentTextBottom = consumed;
-        } else {
+    public void displayCustomText(final String tagElement, final int nixiePositionInRow, final int line) {
+        final Optional<DynamicComponent> lineText = line == 0 ? customTextTop : customTextBottom;
+
+        if (tagElement == null)
             return;
+        if (lineText.filter(d -> d.sameAs(tagElement))
+                .isPresent())
+            return;
+
+        final DynamicComponent component = lineText.orElseGet(DynamicComponent::new);
+        component.displayCustomText(level, worldPosition, tagElement);
+        if (line == 0) {
+            customTextTop = Optional.of(component);
+        } else {
+            customTextBottom = Optional.of(component);
         }
-        Direction right = DoubleOrientedDirections.getLeft(
-                getBlockState().getValue(DoubleOrientedDisplayBlock.FACING),
-                getBlockState().getValue(DoubleOrientedDisplayBlock.ORIENTATION)
-        ).getOpposite();
-        BlockPos rightPos = getBlockPos().relative(right);
-        BlockEntity blockEntity = level.getBlockEntity(rightPos);
-        if (blockEntity instanceof GenericNixieDisplayBlockEntity nextDisplay && areStatesComprableForConnection(getBlockState(), nextDisplay.getBlockState())) {
-            nextDisplay.currentDisplayOption = currentDisplayOption;
-            nextDisplay.applyTextToDisplay(tagElement.isEmpty() ? "" : tagElement.substring(Math.min(consumptionWidth, tagElement.length())), line);
-        }
-        level.setBlockAndUpdate(getBlockPos(), getBlockState().setValue(DoubleOrientedDisplayBlock.LIT,
-                !this.currentTextTop.trim().isEmpty() && (currentDisplayOption.lines == 1 || !this.currentTextTop.trim().isEmpty())));
+        customTextStart = nixiePositionInRow;
         notifyUpdate();
     }
 
     public int seekWidth() {
-        Direction right = DoubleOrientedDirections.getLeft(
-                getBlockState().getValue(DoubleOrientedDisplayBlock.FACING),
-                getBlockState().getValue(DoubleOrientedDisplayBlock.ORIENTATION)
+        final Direction right = DoubleOrientedDirections.getLeft(
+                getBlockState().getValue(GenericNixieDisplayBlock.FACING),
+                getBlockState().getValue(GenericNixieDisplayBlock.ORIENTATION)
         ).getOpposite();
         int characterCount = 0;
         for (int i = 0; i < 100; i++) {
-            BlockPos nextPos = getBlockPos().relative(right, i);
-            BlockEntity blockEntity = level.getBlockEntity(nextPos);
+            final BlockPos nextPos = getBlockPos().relative(right, i);
+            final BlockEntity blockEntity = level.getBlockEntity(nextPos);
             if (!areStatesComprableForConnection(getBlockState(), level.getBlockState(nextPos))) {
                 break;
             }
-            if (blockEntity instanceof GenericNixieDisplayBlockEntity currentWalkNixieDisplay) {
+            if (blockEntity instanceof final GenericNixieDisplayBlockEntity currentWalkNixieDisplay) {
                 characterCount += currentWalkNixieDisplay.calculateDisplayedCharacterWidth();
             } else {
                 CreateBitsnBobs.LOGGER.warn("Found unexpected non-nixie display block entity at {} while seeking width for {}", nextPos, getBlockPos());
@@ -215,8 +298,27 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         return characterCount;
     }
 
+    public void updateDisplayedStrings() {
+        customTextTop.map(DynamicComponent::resolve)
+                .ifPresentOrElse(
+                        fullText -> renderedTextTop = charsOrEmpty(fullText, customTextStart),
+                        () -> renderedTextTop = "");
+        customTextBottom.map(DynamicComponent::resolve)
+                .ifPresentOrElse(
+                        fullText -> renderedTextBottom = charsOrEmpty(fullText, customTextStart),
+                        () -> renderedTextBottom = "");
+    }
+
+    private String charsOrEmpty(final String string, final int index) {
+        final EndClipping endClipping = getEndClipping();
+        return " ".repeat(endClipping.left) +
+                (string.length() <= index ? "   " :
+                        (string + " ".repeat(GenericNixieDisplayBlockEntity.MAX_CHARS_IN_BLOCK))
+                                .substring(index, index + calculateDisplayedCharacterWidth()));
+    }
+
     int calculateDisplayedCharacterWidth() {
-        EndClipping endClipping = getEndClipping();
+        final EndClipping endClipping = getEndClipping();
         return currentDisplayOption.width - endClipping.left - endClipping.right;
     }
 
@@ -229,18 +331,18 @@ public class GenericNixieDisplayBlockEntity extends SmartBlockEntity {
         public final int left;
         public final int right;
 
-        EndClipping(int left, int right) {
+        EndClipping(final int left, final int right) {
             this.left = left;
             this.right = right;
         }
     }
 
     public EndClipping getEndClipping() {
-        if (!(getBlockState().getBlock() instanceof NixieBoardBlock) || currentDisplayOption != ConfigurableDisplayOptions.DOUBLE_CHAR_DOUBLE_LINES) {
+        if (!(getBlockState().getBlock() instanceof NixieBoardBlockNixie) || currentDisplayOption != ConfigurableDisplayOptions.DOUBLE_CHAR_DOUBLE_LINES) {
             return EndClipping.NONE; // Nixie tubes do not have end clipping
         }
-        boolean left = !getBlockState().getValue(NixieBoardBlock.LEFT);
-        boolean right = !getBlockState().getValue(NixieBoardBlock.RIGHT);
+        final boolean left = !getBlockState().getValue(NixieBoardBlockNixie.LEFT);
+        final boolean right = !getBlockState().getValue(NixieBoardBlockNixie.RIGHT);
         if (left && right) {
             return EndClipping.BOTH;
         } else if (left) {