+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/registry/BnbLangEntries.java
@@ -28,6 +28,7 @@ public class BnbLangEntries {
         final String[] entries = {
                 "generator.bits_n_bobs.ponderous_planes", "Ponderflat",
                 "tooltip.bits_n_bobs.new_ponder_notification", "New ponder",
+                "tooltip.bits_n_bobs.chain_drive_placing_hint", "Placing chain drive, create a complete loop to finish.",
                 //Tooltips
                 "block.bits_n_bobs.headlamp.tooltip.summary", "Can be dyed and placed _multiple times in same block_. Useful for trains or fancy signage too!",